put the data in global/data.cs
